Move OpenSearchGrpc to generated file via generate_api.py
Move OpenSearchGrpc out of the auto-generated client/__init__.py into
its own file opensearchpy/client/grpc_client.py, produced by the
generator from utils/templates/grpc_client.

This ensures the gRPC client survives API spec regeneration runs
(nox -rs generate) and follows the project's code generation pattern.

- Add utils/templates/grpc_client template
- Add dump_grpc_client() to utils/generate_api.py
- Create opensearchpy/client/grpc_client.py as generated output
- Remove OpenSearchGrpc from opensearchpy/client/__init__.py
- Export OpenSearchGrpc from opensearchpy/__init__.py
