+++ opensearchpy/client/__init__.py
@@ -3277,96 +3277,3 @@ class OpenSearch(Client):
         )
 
 
-class OpenSearchGrpc(OpenSearch):
-    """
-    OpenSearch client with gRPC transport for bulk document operations.
-
-    Extends the standard OpenSearch client with gRPC channel management.
-    Bulk requests are routed over gRPC for better performance; all other
-    operations fall through to REST automatically.
-
-    Supported parameters:
-        - hosts: REST endpoint(s) for fallback operations
-        - grpc_hosts: gRPC endpoint (required)
-
-    Unsupported parameters (raise NotImplementedError):
-        - use_ssl, verify_certs, ca_certs, client_cert, client_key,
-          ssl_context, ssl_version, ssl_assert_hostname,
-          ssl_assert_fingerprint, ssl_show_warn
-        - http_auth
-
-    Safely ignored (passed to REST fallback only):
-        - http_compress, opaque_id, headers, pool_maxsize
-
-    Usage::
-
-        from opensearchpy.client import OpenSearchGrpc
-
-        client = OpenSearchGrpc(
-            hosts=[{'host': 'localhost', 'port': 9200}],
-            grpc_hosts=[{'host': 'localhost', 'port': 9400}],
-        )
-
-        # Bulk goes over gRPC automatically
-        client.bulk(body=[...])
-
-        # Everything else uses REST
-        client.search(index='my-index', body={'query': {'match_all': {}}})
-
-    :arg hosts: list of REST nodes (same as OpenSearch client).
-    :arg grpc_hosts: list of gRPC nodes, e.g. [{'host': 'localhost', 'port': 9400}].
-    :arg kwargs: all other arguments passed to the OpenSearch client for REST fallback.
-    """
-
-    # Parameters that are not yet supported for gRPC
-    _UNSUPPORTED_TLS_ARGS = (
-        "use_ssl",
-        "verify_certs",
-        "ca_certs",
-        "client_cert",
-        "client_key",
-        "ssl_context",
-        "ssl_version",
-        "ssl_assert_hostname",
-        "ssl_assert_fingerprint",
-        "ssl_show_warn",
-    )
-
-    _UNSUPPORTED_AUTH_ARGS = ("http_auth",)
-
-    def __init__(
-        self,
-        hosts: Any = None,
-        grpc_hosts: Any = None,
-        **kwargs: Any,
-    ) -> None:
-        try:
-            from opensearch_grpc.grpc_transport import GrpcTransport
-        except ImportError as e:
-            raise ImproperlyConfigured(
-                "gRPC dependencies are not installed. "
-                "Install them with: pip install opensearch-py[grpc]"
-            ) from e
-
-        # Check for unsupported TLS parameters
-        for arg in self._UNSUPPORTED_TLS_ARGS:
-            if arg in kwargs and kwargs[arg] is not None and kwargs[arg] is not False:
-                raise NotImplementedError(
-                    f"TLS is not yet supported in the gRPC client. "
-                    f"The '{arg}' parameter cannot be used with OpenSearchGrpc. "
-                    f"This will be supported in a future release."
-                )
-
-        # Check for unsupported auth parameters
-        for arg in self._UNSUPPORTED_AUTH_ARGS:
-            if arg in kwargs and kwargs[arg] is not None:
-                raise NotImplementedError(
-                    f"Authentication is not yet supported in the gRPC client. "
-                    f"The '{arg}' parameter cannot be used with OpenSearchGrpc. "
-                    f"This will be supported in a future release."
-                )
-
-        if grpc_hosts is not None:
-            kwargs["grpc_hosts"] = grpc_hosts
-
-        super().__init__(hosts, transport_class=GrpcTransport, **kwargs)
