+++ test_opensearchpy/test_server/test_grpc/__init__.py
@@ -24,7 +24,7 @@ from urllib.parse import urlparse
 
 import grpc
 
-from opensearchpy.client import OpenSearchGrpc
+from opensearchpy.client.grpc_client import OpenSearchGrpc
 from opensearchpy.helpers.test import OPENSEARCH_URL
 from opensearchpy.helpers.test import OpenSearchTestCase as BaseTestCase
